Updated Juptyer notebook with filter steps and extra chapter on scar clustering
Filtering:
- In percentage calculation: Do not consider site/allele in one cell when it has < 4 reads.
- In allele count calculation: Sequence is considered an allele count if it is present for >55% at a given cell/site/allele.
- In distance clustering: filter out 'bad cells' - cells with 2 or less allele counts.

For visualising scar clusters on tsne:
Every gastruloid is processed independently. Clusters are based on sparse distance matrix, with n clusters as 'sum cells of gastruloid' / 10, resulting in 10 cells per cluster on average. Tsne coordinates are loaded in and this is used to plot all cells from a single gastruloids (in black) and cells from that gastruloid belonging to the same scar cluster on top (red stars). This is done for all gastruloids and all scar clusters.